Fix: 소수형을 받을 수 있도록 변경

+++ src/main/java/com/backend/nutt/dto/request/FormSignUpRequest.java
@@ -11,6 +11,6 @@ public class FormSignUpRequest {
     private String password;
     private String name;
     private String gender;
-    private int height;
-    private int weight;
+    private double height;
+    private double weight;
 }
